Break out Itemset into its own module.
