+++ src/lib/stategraph.rs
@@ -31,18 +31,15 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 use std::collections::HashSet;
-use std::collections::hash_map::{Entry, HashMap};
-use std::hash::BuildHasherDefault;
+use std::collections::hash_map::HashMap;
 
 extern crate bit_vec;
-use self::bit_vec::{BitBlock, BitVec};
-
-extern crate fnv;
-use self::fnv::FnvHasher;
+use self::bit_vec::BitVec;
 
 use StIdx;
 use firsts::Firsts;
-use grammar::{PIdx, Grammar, Symbol, SIdx};
+use grammar::{Grammar, Symbol, SIdx};
+use itemset::Itemset;
 
 // This file creates stategraphs from grammars. Unfortunately there is no perfect guide to how to
 // do this that I know of -- certainly not one that talks about sensible ways to arrange data and
@@ -58,143 +55,7 @@ use grammar::{PIdx, Grammar, Symbol, SIdx};
 //   Measuring and extending LR(1) parser generation
 //     Xin Chen, PhD thesis, University of Hawaii, 2009
 
-
-/// The type of "context" (also known as "lookaheads")
-pub type Ctx = BitVec;
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct Itemset {
-    pub items: HashMap<(PIdx, SIdx), Ctx, BuildHasherDefault<FnvHasher>>
-}
-
 impl Itemset {
-    /// Create a blank Itemset.
-    fn new(_: &Grammar) -> Itemset {
-        Itemset {items: HashMap::with_hasher(BuildHasherDefault::<FnvHasher>::default())}
-    }
-
-    /// Add an item `(prod, dot)` with context `ctx` to this itemset. Returns true if this led to
-    /// any changes in the itemset.
-    fn add(&mut self, prod: PIdx, dot: SIdx, ctx: &Ctx) -> bool {
-        let entry = self.items.entry((prod, dot));
-        match entry {
-            Entry::Occupied(mut e) => {
-                e.get_mut().union(ctx)
-            }
-            Entry::Vacant(e)       => {
-                e.insert(ctx.clone());
-                true
-            }
-        }
-    }
-
-    /// Create a new itemset which is a closed version of `self`.
-    fn close(&self, grm: &Grammar, firsts: &Firsts) -> Itemset {
-        // This function can be seen as a merger of getClosure and getContext from Chen's
-        // dissertation.
-
-        let mut new_is = self.clone(); // The new itemset we're building up
-
-        // In a typical description of this algorithm, one would have a todo set which contains
-        // pairs (prod_i, dot). Unfortunately this is a slow way of doing things. Searching the set
-        // for the next item and removing it is slow; and, since we don't know how many potential
-        // dots there are in a production, the set is of potentially unbounded size, so we can end
-        // up resizing memory. Since this function is the most expensive in the table generation,
-        // using a HashSet (which is the "obvious" solution) is slow.
-        //
-        // However, we can reduce these costs through two observations:
-        //   1) The initial todo set is populated with (prod_i, dot) pairs that all come from
-        //      self.items.keys(). There's no point copying these into a todo list.
-        //   2) All subsequent todo items are of the form (prod_off, 0). Since the dot in these
-        //      cases is always 0, we don't need to store pairs: simply knowing which prod_off's we
-        //      need to look at is sufficient. We can represent these with a fixed-size bitfield.
-        // All we need to do is first iterate through the items in 1 and, when it's exhausted,
-        // continually iterate over the bitfield from 2 until no new items have been added.
-
-        let mut keys_iter = self.items.keys(); // The initial todo list
-        type BitVecBitSize = u32; // As of 0.4.3, BitVec only supports u32 blocks
-        let mut zero_todos = BitVec::<BitVecBitSize>::from_elem(grm.prods_len, false); // Subsequent todos
-        let mut new_ctx = BitVec::from_elem(grm.terms_len, false);
-        loop {
-            let prod_i;
-            let dot;
-            // Find the next todo item or, if there are none, break the loop. First of all we try
-            // pumping keys_iter for its next value. If there is none (i.e. we've exhausted that
-            // part of the todo set), we iterate over zero_todos.
-            match keys_iter.next() {
-                Some(&(x, y)) => {
-                    prod_i = x;
-                    dot = y;
-                }
-                None => {
-                    // Quickly iterate over the BitVec looking for the first non-zero word.
-                    match zero_todos.blocks().enumerate().filter(|&(_, b)| b != 0).next() {
-                        Some((b_off, b)) => {
-                            // prod_i is the block offset plus the index of the first bit set.
-                            prod_i = PIdx::from(b_off * BitVecBitSize::bits() + (b.trailing_zeros() as usize))
-                        },
-                        None => break
-                    }
-                    dot = SIdx::from(0);
-                    zero_todos.set(prod_i.into(), false);
-                }
-            }
-            let prod = grm.prod(prod_i).unwrap();
-            if dot == SIdx::from(prod.len()) { continue; }
-            if let Symbol::Nonterminal(nonterm_i) = prod[usize::from(dot)] {
-                // This if statement is, in essence, a fast version of what's called getContext in
-                // Chen's dissertation, folding in getTHeads at the same time. The particular
-                // formulation here is based as much on
-                // http://binarysculpting.com/2012/02/04/computing-lr1-closure/ as it is any of the
-                // several versions of getTHeads in Chen's dissertation. Nevertheless, the mapping
-                // between the two different formulations is fairly straight-forward.
-                new_ctx.clear();
-                let mut nullable = true;
-                for sym in prod.iter().skip(usize::from(dot) + 1) {
-                    match sym {
-                        &Symbol::Terminal(term_j) => {
-                            new_ctx.set(usize::from(term_j), true);
-                            nullable = false;
-                            break;
-                        },
-                        &Symbol::Nonterminal(nonterm_j) => {
-                            new_ctx.union(firsts.prod_firsts(nonterm_j));
-                            if !firsts.is_epsilon_set(nonterm_j) {
-                                nullable = false;
-                                break;
-                            }
-                        }
-                    }
-                }
-                if nullable {
-                    new_ctx.union(&new_is.items[&(prod_i, dot)]);
-                }
-
-                for ref_prod_i in grm.nonterm_to_prods(nonterm_i).unwrap().iter() {
-                    if new_is.add(*ref_prod_i, SIdx::from(0), &new_ctx) {
-                        zero_todos.set(usize::from(*ref_prod_i), true);
-                    }
-                }
-            }
-        }
-        new_is
-    }
-
-    /// Create a new Itemset based on calculating the goto of 'sym' on the current Itemset.
-    fn goto(&self, grm: &Grammar, sym: &Symbol) -> Itemset {
-        // This is called 'transition' in Chen's dissertation, though note that the definition
-        // therein appears to get the dot in the input/output the wrong way around.
-        let mut newis = Itemset::new(grm);
-        for (&(prod_i, dot), ctx) in &self.items {
-            let prod = grm.prod(prod_i).unwrap();
-            if dot == SIdx::from(prod.len()) { continue; }
-            if sym == &prod[usize::from(dot)] {
-                newis.add(prod_i, SIdx::from(usize::from(dot) + 1), ctx);
-            }
-        }
-        newis
-    }
-
     /// Return true if `other` is weakly compatible with `self`.
     fn weakly_compatible(&self, other: &Itemset) -> bool {
         // The weakly compatible check is one of the three core parts of Pager's algorithm
@@ -510,12 +371,39 @@ impl StateGraph {
 }
 
 #[cfg(test)]
+pub fn state_exists(grm: &Grammar, is: &Itemset, nt: &str, prod_off: usize, dot: usize, la:
+                    Vec<&str>) {
+    use grammar::TIdx;
+
+    let ab_prod_off = grm.nonterm_to_prods(grm.nonterminal_off(nt)).unwrap()[prod_off];
+    let ctx = &is.items[&(ab_prod_off, SIdx::from(dot))];
+    for i in 0..grm.terms_len {
+        let bit = ctx[i];
+        let mut found = false;
+        for t in la.iter() {
+            if grm.terminal_off(t) == TIdx::from(i) {
+                if !bit {
+                    panic!("bit for terminal {}, dot {} is not set in production {} of {} when it should be",
+                           t, dot, prod_off, nt);
+                }
+                found = true;
+                break;
+            }
+        }
+        if !found && bit {
+            panic!("bit for terminal {}, dot {} is set in production {} of {} when it shouldn't be",
+                   grm.term_name(TIdx::from(i)).unwrap(), dot, prod_off, nt);
+        }
+    }
+}
+
+#[cfg(test)]
 mod test {
     extern crate bit_vec;
     use self::bit_vec::BitVec;
 
-    use super::{bitvec_intersect, Itemset, Firsts, StateGraph};
-    use grammar::{Grammar, SIdx, Symbol, TIdx};
+    use super::{bitvec_intersect, state_exists, StateGraph};
+    use grammar::{Grammar, Symbol};
     use yacc_parser::parse_yacc;
 
     #[test]
@@ -543,101 +431,6 @@ mod test {
         assert!(bitvec_intersect(&b1, &b2));
     }
 
-    fn state_exists(grm: &Grammar, is: &Itemset, nt: &str, prod_off: usize, dot: usize, la:
-                        Vec<&str>) {
-        let ab_prod_off = grm.nonterm_to_prods(grm.nonterminal_off(nt)).unwrap()[prod_off];
-        let ctx = &is.items[&(ab_prod_off, SIdx::from(dot))];
-        for i in 0..grm.terms_len {
-            let bit = ctx[i];
-            let mut found = false;
-            for t in la.iter() {
-                if grm.terminal_off(t) == TIdx::from(i) {
-                    if !bit {
-                        panic!("bit for terminal {}, dot {} is not set in production {} of {} when it should be",
-                               t, dot, prod_off, nt);
-                    }
-                    found = true;
-                    break;
-                }
-            }
-            if !found && bit {
-                panic!("bit for terminal {}, dot {} is set in production {} of {} when it shouldn't be",
-                       grm.term_name(TIdx::from(i)).unwrap(), dot, prod_off, nt);
-            }
-        }
-    }
-
-    fn num_active_states(is: &Itemset) -> usize {
-        is.items.len()
-    }
-
-    #[test]
-    fn test_dragon_grammar() {
-        // From http://binarysculpting.com/2012/02/04/computing-lr1-closure/
-        let grm = Grammar::new(&parse_yacc(&"
-          %start S
-          %%
-          S: L '=' R | R;
-          L: '*' R | 'id';
-          R: L;
-          ".to_string()).unwrap());
-        let firsts = Firsts::new(&grm);
-
-        let mut is = Itemset::new(&grm);
-        let mut la = BitVec::from_elem(grm.terms_len, false);
-        la.set(usize::from(grm.terminal_off("$")), true);
-        is.add(grm.nonterm_to_prods(grm.nonterminal_off("^")).unwrap()[0], SIdx::from(0), &la);
-        let cls_is = is.close(&grm, &firsts);
-        println!("{:?}", cls_is);
-        assert_eq!(num_active_states(&cls_is), 6);
-        state_exists(&grm, &cls_is, "^", 0, 0, vec!["$"]);
-        state_exists(&grm, &cls_is, "S", 0, 0, vec!["$"]);
-        state_exists(&grm, &cls_is, "S", 1, 0, vec!["$"]);
-        state_exists(&grm, &cls_is, "L", 0, 0, vec!["$", "="]);
-        state_exists(&grm, &cls_is, "L", 1, 0, vec!["$", "="]);
-        state_exists(&grm, &cls_is, "R", 0, 0, vec!["$"]);
-    }
-
-    fn eco_grammar() -> Grammar {
-        let ast = parse_yacc(&"
-          %start S
-          %token a b c d f
-          %%
-          S: S 'b' | 'b' A 'a' | 'a';
-          A: 'a' S 'c' | 'a' | 'a' S 'b';
-          B: A S;
-          C: D A;
-          D: 'd' | ;
-          F: C D 'f';
-          ".to_string()).unwrap();
-        Grammar::new(&ast)
-    }
-
-    #[test]
-    fn test_closure1_ecogrm() {
-        let grm = eco_grammar();
-        let firsts = Firsts::new(&grm);
-
-        let mut is = Itemset::new(&grm);
-        let mut la = BitVec::from_elem(grm.terms_len, false);
-        la.set(usize::from(grm.terminal_off("$")), true);
-        is.add(grm.nonterm_to_prods(grm.nonterminal_off("^")).unwrap()[0], SIdx::from(0), &la);
-        let mut cls_is = is.close(&grm, &firsts);
-
-        state_exists(&grm, &cls_is, "^", 0, 0, vec!["$"]);
-        state_exists(&grm, &cls_is, "S", 0, 0, vec!["b", "$"]);
-        state_exists(&grm, &cls_is, "S", 1, 0, vec!["b", "$"]);
-        state_exists(&grm, &cls_is, "S", 2, 0, vec!["b", "$"]);
-
-        is = Itemset::new(&grm);
-        is.add(grm.nonterm_to_prods(grm.nonterminal_off("F")).unwrap()[0], SIdx::from(0), &la);
-        cls_is = is.close(&grm, &firsts);
-        state_exists(&grm, &cls_is, "F", 0, 0, vec!["$"]);
-        state_exists(&grm, &cls_is, "C", 0, 0, vec!["d", "f"]);
-        state_exists(&grm, &cls_is, "D", 0, 0, vec!["a"]);
-        state_exists(&grm, &cls_is, "D", 1, 0, vec!["a"]);
-    }
-
     // GrammarAST from 'LR(k) Analyse fuer Pragmatiker'
     // Z : S
     // S : Sb
@@ -656,65 +449,6 @@ mod test {
     }
 
     #[test]
-    fn test_closure1_grm3() {
-        let grm = grammar3();
-        let firsts = Firsts::new(&grm);
-
-        let mut is = Itemset::new(&grm);
-        let mut la = BitVec::from_elem(grm.terms_len, false);
-        la.set(usize::from(grm.terminal_off("$")), true);
-        is.add(grm.nonterm_to_prods(grm.nonterminal_off("^")).unwrap()[0], SIdx::from(0), &la);
-        let mut cls_is = is.close(&grm, &firsts);
-
-        state_exists(&grm, &cls_is, "^", 0, 0, vec!["$"]);
-        state_exists(&grm, &cls_is, "S", 0, 0, vec!["b", "$"]);
-        state_exists(&grm, &cls_is, "S", 1, 0, vec!["b", "$"]);
-
-        is = Itemset::new(&grm);
-        la = BitVec::from_elem(grm.terms_len, false);
-        la.set(usize::from(grm.terminal_off("b")), true);
-        la.set(usize::from(grm.terminal_off("$")), true);
-        is.add(grm.nonterm_to_prods(grm.nonterminal_off("S")).unwrap()[1], SIdx::from(1), &la);
-        cls_is = is.close(&grm, &firsts);
-        state_exists(&grm, &cls_is, "A", 0, 0, vec!["a"]);
-        state_exists(&grm, &cls_is, "A", 1, 0, vec!["a"]);
-        state_exists(&grm, &cls_is, "A", 2, 0, vec!["a"]);
-
-        is = Itemset::new(&grm);
-        la = BitVec::from_elem(grm.terms_len, false);
-        la.set(usize::from(grm.terminal_off("a")), true);
-        is.add(grm.nonterm_to_prods(grm.nonterminal_off("A")).unwrap()[0], SIdx::from(1), &la);
-        cls_is = is.close(&grm, &firsts);
-        state_exists(&grm, &cls_is, "S", 0, 0, vec!["b", "c"]);
-        state_exists(&grm, &cls_is, "S", 1, 0, vec!["b", "c"]);
-    }
-
-    #[test]
-    fn test_goto1() {
-        let grm = grammar3();
-        let firsts = Firsts::new(&grm);
-
-        let mut is = Itemset::new(&grm);
-        let mut la = BitVec::from_elem(grm.terms_len, false);
-        la.set(usize::from(grm.terminal_off("$")), true);
-        is.add(grm.nonterm_to_prods(grm.nonterminal_off("^")).unwrap()[0], SIdx::from(0), &la);
-        let cls_is = is.close(&grm, &firsts);
-
-        let goto1 = cls_is.goto(&grm, &Symbol::Nonterminal(grm.nonterminal_off("S")));
-        state_exists(&grm, &goto1, "^", 0, 1, vec!["$"]);
-        state_exists(&grm, &goto1, "S", 0, 1, vec!["$", "b"]);
-
-        // follow 'b' from start set
-        let goto2 = cls_is.goto(&grm, &Symbol::Terminal(grm.terminal_off("b")));
-        state_exists(&grm, &goto2, "S", 1, 1, vec!["$", "b"]);
-
-        // continue by following 'a' from last goto, after it's been closed
-        let goto3 = goto2.close(&grm, &firsts).goto(&grm, &Symbol::Terminal(grm.terminal_off("a")));
-        state_exists(&grm, &goto3, "A", 1, 1, vec!["a"]);
-        state_exists(&grm, &goto3, "A", 2, 1, vec!["a"]);
-    }
-
-    #[test]
     fn test_stategraph() {
         let grm = grammar3();
         let sg = StateGraph::new(&grm);
@@ -722,39 +456,39 @@ mod test {
         assert_eq!(sg.states.len(), 10);
         assert_eq!(sg.edges.iter().fold(0, |a, e| a + e.len()), 10);
 
-        assert_eq!(num_active_states(&sg.states[0]), 3);
+        assert_eq!(sg.states[0].items.len(), 3);
         state_exists(&grm, &sg.states[0], "^", 0, 0, vec!["$"]);
         state_exists(&grm, &sg.states[0], "S", 0, 0, vec!["$", "b"]);
         state_exists(&grm, &sg.states[0], "S", 1, 0, vec!["$", "b"]);
 
         let s1 = sg.edges[0][&Symbol::Nonterminal(grm.nonterminal_off("S"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s1)]), 2);
+        assert_eq!(sg.states[usize::from(s1)].items.len(), 2);
         state_exists(&grm, &sg.states[usize::from(s1)], "^", 0, 1, vec!["$"]);
         state_exists(&grm, &sg.states[usize::from(s1)], "S", 0, 1, vec!["$", "b"]);
 
         let s2 = sg.edges[usize::from(s1)][&Symbol::Terminal(grm.terminal_off("b"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s2)]), 1);
+        assert_eq!(sg.states[usize::from(s2)].items.len(), 1);
         state_exists(&grm, &sg.states[usize::from(s2)], "S", 0, 2, vec!["$", "b"]);
 
         let s3 = sg.edges[0][&Symbol::Terminal(grm.terminal_off("b"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s3)]), 4);
+        assert_eq!(sg.states[usize::from(s3)].items.len(), 4);
         state_exists(&grm, &sg.states[usize::from(s3)], "S", 1, 1, vec!["$", "b", "c"]);
         state_exists(&grm, &sg.states[usize::from(s3)], "A", 0, 0, vec!["a"]);
         state_exists(&grm, &sg.states[usize::from(s3)], "A", 1, 0, vec!["a"]);
         state_exists(&grm, &sg.states[usize::from(s3)], "A", 2, 0, vec!["a"]);
 
         let s4 = sg.edges[usize::from(s3)][&Symbol::Nonterminal(grm.nonterminal_off("A"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s4)]), 1);
+        assert_eq!(sg.states[usize::from(s4)].items.len(), 1);
         state_exists(&grm, &sg.states[usize::from(s4)], "S", 1, 2, vec!["$", "b", "c"]);
 
         let s5 = sg.edges[usize::from(s4)][&Symbol::Terminal(grm.terminal_off("a"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s5)]), 1);
+        assert_eq!(sg.states[usize::from(s5)].items.len(), 1);
         state_exists(&grm, &sg.states[usize::from(s5)], "S", 1, 3, vec!["$", "b", "c"]);
 
         let s6 = sg.edges[usize::from(s3)][&Symbol::Terminal(grm.terminal_off("a"))];
         // result from merging 10 into 3
         assert_eq!(s3, sg.edges[usize::from(s6)][&Symbol::Terminal(grm.terminal_off("b"))]);
-        assert_eq!(num_active_states(&sg.states[usize::from(s6)]), 5);
+        assert_eq!(sg.states[usize::from(s6)].items.len(), 5);
         state_exists(&grm, &sg.states[usize::from(s6)], "A", 0, 1, vec!["a"]);
         state_exists(&grm, &sg.states[usize::from(s6)], "A", 1, 1, vec!["a"]);
         state_exists(&grm, &sg.states[usize::from(s6)], "A", 2, 1, vec!["a"]);
@@ -762,17 +496,17 @@ mod test {
         state_exists(&grm, &sg.states[usize::from(s6)], "S", 1, 0, vec!["b", "c"]);
 
         let s7 = sg.edges[usize::from(s6)][&Symbol::Nonterminal(grm.nonterminal_off("S"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s7)]), 3);
+        assert_eq!(sg.states[usize::from(s7)].items.len(), 3);
         state_exists(&grm, &sg.states[usize::from(s7)], "A", 0, 2, vec!["a"]);
         state_exists(&grm, &sg.states[usize::from(s7)], "A", 2, 2, vec!["a"]);
         state_exists(&grm, &sg.states[usize::from(s7)], "S", 0, 1, vec!["b", "c"]);
 
         let s8 = sg.edges[usize::from(s7)][&Symbol::Terminal(grm.terminal_off("c"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s8)]), 1);
+        assert_eq!(sg.states[usize::from(s8)].items.len(), 1);
         state_exists(&grm, &sg.states[usize::from(s8)], "A", 0, 3, vec!["a"]);
 
         let s9 = sg.edges[usize::from(s7)][&Symbol::Terminal(grm.terminal_off("b"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s9)]), 2);
+        assert_eq!(sg.states[usize::from(s9)].items.len(), 2);
         state_exists(&grm, &sg.states[usize::from(s9)], "A", 2, 3, vec!["a"]);
         state_exists(&grm, &sg.states[usize::from(s9)], "S", 0, 2, vec!["b", "c"]);
     }
@@ -798,7 +532,7 @@ mod test {
         assert_eq!(sg.edges.iter().fold(0, |a, e| a + e.len()), 27);
 
         // State 0
-        assert_eq!(num_active_states(&sg.states[0]), 7);
+        assert_eq!(sg.states[0].items.len(), 7);
         state_exists(&grm, &sg.states[0], "^", 0, 0, vec!["$"]);
         state_exists(&grm, &sg.states[0], "X", 0, 0, vec!["$"]);
         state_exists(&grm, &sg.states[0], "X", 1, 0, vec!["$"]);
@@ -808,7 +542,7 @@ mod test {
         state_exists(&grm, &sg.states[0], "X", 5, 0, vec!["$"]);
 
         let s1 = sg.edges[0][&Symbol::Terminal(grm.terminal_off("a"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s1)]), 7);
+        assert_eq!(sg.states[usize::from(s1)].items.len(), 7);
         state_exists(&grm, &sg.states[usize::from(s1)], "X", 0, 1, vec!["a", "d", "e", "$"]);
         state_exists(&grm, &sg.states[usize::from(s1)], "X", 1, 1, vec!["a", "d", "e", "$"]);
         state_exists(&grm, &sg.states[usize::from(s1)], "X", 2, 1, vec!["a", "d", "e", "$"]);
@@ -818,7 +552,7 @@ mod test {
         state_exists(&grm, &sg.states[usize::from(s1)], "T", 0, 0, vec!["a", "d", "e", "$"]);
 
         let s7 = sg.edges[0][&Symbol::Terminal(grm.terminal_off("b"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s7)]), 7);
+        assert_eq!(sg.states[usize::from(s7)].items.len(), 7);
         state_exists(&grm, &sg.states[usize::from(s7)], "X", 3, 1, vec!["a", "d", "e", "$"]);
         state_exists(&grm, &sg.states[usize::from(s7)], "X", 4, 1, vec!["a", "d", "e", "$"]);
         state_exists(&grm, &sg.states[usize::from(s7)], "X", 5, 1, vec!["a", "d", "e", "$"]);
@@ -828,7 +562,7 @@ mod test {
         state_exists(&grm, &sg.states[usize::from(s1)], "T", 0, 0, vec!["a", "d", "e", "$"]);
 
         let s4 = sg.edges[usize::from(s1)][&Symbol::Terminal(grm.terminal_off("u"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s4)]), 8);
+        assert_eq!(sg.states[usize::from(s4)].items.len(), 8);
         assert_eq!(s4, sg.edges[usize::from(s7)][&Symbol::Terminal(grm.terminal_off("u"))]);
         state_exists(&grm, &sg.states[usize::from(s4)], "Y", 1, 1, vec!["d", "e"]);
         state_exists(&grm, &sg.states[usize::from(s4)], "T", 0, 1, vec!["a", "d", "e", "$"]);
@@ -843,34 +577,34 @@ mod test {
         assert_eq!(s7, sg.edges[usize::from(s4)][&Symbol::Terminal(grm.terminal_off("b"))]);
 
         let s2 = sg.edges[usize::from(s1)][&Symbol::Terminal(grm.terminal_off("t"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s2)]), 3);
+        assert_eq!(sg.states[usize::from(s2)].items.len(), 3);
         state_exists(&grm, &sg.states[usize::from(s2)], "Y", 0, 1, vec!["d"]);
         state_exists(&grm, &sg.states[usize::from(s2)], "Z", 0, 1, vec!["c"]);
         state_exists(&grm, &sg.states[usize::from(s2)], "W", 0, 0, vec!["d"]);
 
         let s3 = sg.edges[usize::from(s2)][&Symbol::Terminal(grm.terminal_off("u"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s3)]), 3);
+        assert_eq!(sg.states[usize::from(s3)].items.len(), 3);
         state_exists(&grm, &sg.states[usize::from(s3)], "Z", 0, 2, vec!["c"]);
         state_exists(&grm, &sg.states[usize::from(s3)], "W", 0, 1, vec!["d"]);
         state_exists(&grm, &sg.states[usize::from(s3)], "V", 0, 0, vec!["d"]);
 
         let s5 = sg.edges[usize::from(s4)][&Symbol::Nonterminal(grm.nonterminal_off("X"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s5)]), 2);
+        assert_eq!(sg.states[usize::from(s5)].items.len(), 2);
         state_exists(&grm, &sg.states[usize::from(s5)], "Y", 1, 2, vec!["d", "e"]);
         state_exists(&grm, &sg.states[usize::from(s5)], "T", 0, 2, vec!["a", "d", "e", "$"]);
 
         let s6 = sg.edges[usize::from(s5)][&Symbol::Terminal(grm.terminal_off("a"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s6)]), 1);
+        assert_eq!(sg.states[usize::from(s6)].items.len(), 1);
         state_exists(&grm, &sg.states[usize::from(s6)], "T", 0, 3, vec!["a", "d", "e", "$"]);
 
         let s8 = sg.edges[usize::from(s7)][&Symbol::Terminal(grm.terminal_off("t"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s8)]), 3);
+        assert_eq!(sg.states[usize::from(s8)].items.len(), 3);
         state_exists(&grm, &sg.states[usize::from(s8)], "Y", 0, 1, vec!["e"]);
         state_exists(&grm, &sg.states[usize::from(s8)], "Z", 0, 1, vec!["d"]);
         state_exists(&grm, &sg.states[usize::from(s8)], "W", 0, 0, vec!["e"]);
 
         let s9 = sg.edges[usize::from(s8)][&Symbol::Terminal(grm.terminal_off("u"))];
-        assert_eq!(num_active_states(&sg.states[usize::from(s9)]), 3);
+        assert_eq!(sg.states[usize::from(s9)].items.len(), 3);
         state_exists(&grm, &sg.states[usize::from(s9)], "Z", 0, 2, vec!["d"]);
         state_exists(&grm, &sg.states[usize::from(s9)], "W", 0, 1, vec!["e"]);
         state_exists(&grm, &sg.states[usize::from(s3)], "V", 0, 0, vec!["d"]);
